kvm ppc inst

+++ arch/powerpc/include/asm/disassemble.h
@@ -10,73 +10,74 @@
 #define __ASM_PPC_DISASSEMBLE_H__
 
 #include <linux/types.h>
+#include <asm/inst.h>
 
-static inline unsigned int get_op(u32 inst)
+static inline unsigned int get_op(unsigned long inst)
 {
 	return inst >> 26;
 }
 
-static inline unsigned int get_xop(u32 inst)
+static inline unsigned int get_xop(unsigned long inst)
 {
 	return (inst >> 1) & 0x3ff;
 }
 
-static inline unsigned int get_sprn(u32 inst)
+static inline unsigned int get_sprn(unsigned long inst)
 {
 	return ((inst >> 16) & 0x1f) | ((inst >> 6) & 0x3e0);
 }
 
-static inline unsigned int get_dcrn(u32 inst)
+static inline unsigned int get_dcrn(unsigned long inst)
 {
 	return ((inst >> 16) & 0x1f) | ((inst >> 6) & 0x3e0);
 }
 
-static inline unsigned int get_tmrn(u32 inst)
+static inline unsigned int get_tmrn(unsigned long inst)
 {
 	return ((inst >> 16) & 0x1f) | ((inst >> 6) & 0x3e0);
 }
 
-static inline unsigned int get_rt(u32 inst)
+static inline unsigned int get_rt(unsigned long inst)
 {
 	return (inst >> 21) & 0x1f;
 }
 
-static inline unsigned int get_rs(u32 inst)
+static inline unsigned int get_rs(unsigned long inst)
 {
 	return (inst >> 21) & 0x1f;
 }
 
-static inline unsigned int get_ra(u32 inst)
+static inline unsigned int get_ra(unsigned long inst)
 {
 	return (inst >> 16) & 0x1f;
 }
 
-static inline unsigned int get_rb(u32 inst)
+static inline unsigned int get_rb(unsigned long inst)
 {
 	return (inst >> 11) & 0x1f;
 }
 
-static inline unsigned int get_rc(u32 inst)
+static inline unsigned int get_rc(unsigned long inst)
 {
 	return inst & 0x1;
 }
 
-static inline unsigned int get_ws(u32 inst)
+static inline unsigned int get_ws(unsigned long inst)
 {
 	return (inst >> 11) & 0x1f;
 }
 
-static inline unsigned int get_d(u32 inst)
+static inline unsigned int get_d(unsigned long inst)
 {
 	return inst & 0xffff;
 }
 
-static inline unsigned int get_oc(u32 inst)
+static inline unsigned int get_oc(unsigned long inst)
 {
 	return (inst >> 11) & 0x7fff;
 }
 
-static inline unsigned int get_tx_or_sx(u32 inst)
+static inline unsigned int get_tx_or_sx(unsigned long inst)
 {
 	return (inst) & 0x1;
 }
@@ -87,7 +88,7 @@ static inline unsigned int get_tx_or_sx(u32 inst)
 /*
  * Create a DSISR value from the instruction
  */
-static inline unsigned make_dsisr(unsigned instr)
+static inline unsigned make_dsisr(unsigned long instr)
 {
 	unsigned dsisr;
